LayoutTests/imported/w3c/web-platform-tests/webxr/dom-overlay/ar_dom_overlay_hit_test.https-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: XRRay


Harness Error (FAIL), message = ReferenceError: Can't find variable: XRRay
